requirements.txt
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
cloudscraper==1.2.71
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.66.0
idna==3.8
numpy
pandas
proto-plus
protobuf
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
